gcp_bigtable_nosql_de/pipelines/requirements.txt
# Core Google Cloud Libraries
google-cloud-bigtable>=2.19.0
google-cloud-logging>=3.8.0
google-cloud-monitoring>=2.16.0
google-cloud-pubsub>=2.18.0

# Data Processing
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.24.0

# Data Quality
great-expectations>=0.18.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
retrying>=1.3.4

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
pylint>=2.17.0
mypy>=1.5.0

# Monitoring and Alerting
slack-sdk>=3.21.0
